refactor: move metadata to metadata folder
