+++ providers/brevo/connector.go
@@ -6,20 +6,8 @@ import (
 	"github.com/amp-labs/connectors/common"
 	"github.com/amp-labs/connectors/internal/components"
 	"github.com/amp-labs/connectors/internal/components/schema"
-	"github.com/amp-labs/connectors/internal/staticschema"
 	"github.com/amp-labs/connectors/providers"
-	"github.com/amp-labs/connectors/tools/fileconv"
-	"github.com/amp-labs/connectors/tools/scrapper"
-)
-
-// nolint:gochecknoglobals
-var (
-	//go:embed schemas.json
-	schemaContent []byte
-	FileManager   = scrapper.NewMetadataFileManager[staticschema.FieldMetadataMapV2](
-		schemaContent, fileconv.NewSiblingFileLocator())
-
-	schemas = FileManager.MustLoadSchemas()
+	"github.com/amp-labs/connectors/providers/brevo/metadata"
 )
 
 type Connector struct {
@@ -46,7 +34,7 @@ func constructor(base *components.Connector) (*Connector, error) {
 	connector := &Connector{Connector: base}
 
 	// Set the metadata provider for the connector
-	connector.SchemaProvider = schema.NewOpenAPISchemaProvider(connector.ProviderContext.Module(), schemas)
+	connector.SchemaProvider = schema.NewOpenAPISchemaProvider(connector.ProviderContext.Module(), metadata.Schemas)
 
 	return connector, nil
 }
